fix-api(valores monetarios): alterando tipo de algumas colunas de valores monetarios de float para decimal

+++ database/migrations/2021_04_06_175408_alter_table_pessoa_assinatura_for_pj.php
@@ -36,13 +36,13 @@ class AlterTablePessoaAssinaturaForPj extends Migration
                 ->unsigned()
                 ->constrained('cad_porte_empresa', 'id_porte_empresa');
 
-            $table->float('rendimento_mensal')
+            $table->decimal('rendimento_mensal', 12, 2)
                 ->nullable();
 
-            $table->float('faturamento_anual')
+            $table->decimal('faturamento_anual', 12, 2)
                 ->nullable();
 
-            $table->float('capital_social')
+            $table->decimal('capital_social', 12, 2)
                 ->nullable();
 
             $table->integer('ano_faturamento')
